Fixes webhook creation on news channels
