+++ src/model/channel/channel_id.rs
@@ -6,6 +6,12 @@ use std::sync::Arc;
 use futures::stream::Stream;
 #[cfg(feature = "model")]
 use serde_json::json;
+#[cfg(feature = "model")]
+use tokio::{fs::File, io::AsyncReadExt};
+#[cfg(feature = "model")]
+use reqwest::Url;
+#[cfg(feature = "model")]
+use bytes::buf::Buf;
 
 #[cfg(feature = "model")]
 use crate::builder::{CreateInvite, CreateMessage, EditChannel, EditMessage, GetMessages};
@@ -851,6 +857,78 @@ impl ChannelId {
     #[inline]
     pub async fn webhooks(self, http: impl AsRef<Http>) -> Result<Vec<Webhook>> {
         http.as_ref().get_channel_webhooks(self.0).await
+    }
+
+    /// Creates a webhook with only a name.
+    ///
+    /// # Errors
+    ///
+    /// Returns a [`Error::Http`] if the current user lacks permission.
+    pub async fn create_webhook(
+        &self,
+        http: impl AsRef<Http>,
+        name: impl std::fmt::Display,
+    ) -> Result<Webhook> {
+        let map = serde_json::json!({
+            "name": name.to_string(),
+        });
+
+        http.as_ref().create_webhook(self.0, &map).await
+    }
+
+    /// Creates a webhook with a name and an avatar.
+    ///
+    /// # Errors
+    ///
+    /// In addition to the reasons [`Self::create_webhook`] may return an [`Error::Http`],
+    /// if the image is too large.
+    pub async fn create_webhook_with_avatar<'a>(
+        &self,
+        http: impl AsRef<Http>,
+        name: impl std::fmt::Display,
+        avatar: impl Into<AttachmentType<'a>>,
+    ) -> Result<Webhook> {
+        let name = name.to_string();
+        let avatar = avatar.into();
+
+        let avatar = match avatar {
+            AttachmentType::Bytes {
+                data,
+                filename: _,
+            } => "data:image/png;base64,".to_string() + &base64::encode(&data.into_owned()),
+            AttachmentType::File {
+                file,
+                filename: _,
+            } => {
+                let mut buf = Vec::new();
+                file.try_clone().await?.read_to_end(&mut buf).await?;
+
+                "data:image/png;base64,".to_string() + &base64::encode(&buf)
+            },
+            AttachmentType::Path(path) => {
+                let mut file = File::open(path).await?;
+                let mut buf = vec![];
+                file.read_to_end(&mut buf).await?;
+
+                "data:image/png;base64,".to_string() + &base64::encode(&buf)
+            },
+            AttachmentType::Image(url) => {
+                let url = Url::parse(url).map_err(|_| Error::Url(url.to_string()))?;
+                let response = http.as_ref().client.get(url).send().await?;
+                let mut bytes = response.bytes().await?;
+                let mut picture: Vec<u8> = vec![0; bytes.len()];
+                bytes.copy_to_slice(&mut picture[..]);
+
+                "data:image/png;base64,".to_string() + &base64::encode(&picture)
+            },
+        };
+
+        let map = serde_json::json!({
+            "name": name,
+            "avatar": avatar
+        });
+
+        http.as_ref().create_webhook(self.0, &map).await
     }
 
     /// Returns a future that will await one message sent in this channel.
